Create RegLog-valid database tables with DBI

Create RegLog-valid database tables with DBI

Usage

DBI_tables_create(
  conn,
  account_name = "account",
  reset_code_name = "reset_code",
  use_log = FALSE,
  log_name = "logs",
  user_data = NULL,
  hash_passwords = FALSE,
  verbose = TRUE
)

Arguments

conn

DBI connection object

account_name

Name of the table for storing user accounts credentials. Defaults to 'account'. Mandatory table.

reset_code_name

Name of the table for storing generated password reset codes. Defaults to 'reset_code'. Mandatory table.

use_log

Should the table for keeping RegLogServer logs be also created? Defaults to FALSE

log_name

Name of the table for storing logs from RegLogServer object. Used only if use_log = TRUE. Defaults to logs

user_data

If you wish to import existing user database, you can input data.frame with that table in this argument. It should contain columns: username, password, email. Defaults to NULL.

hash_passwords

If you are importing table of users upon tables creation, you can also specify if the password should be hashed using scrypt::hashPassword. Defaults to FALSE. If you have unhashed passwords in imported table, set this option to TRUE.

verbose

Boolean specific if the actions made by function should be printed back to the console. Defaults to TRUE.

Value

List with results of the creation

Details

Currently, the function is tested and working correctly for SQLite, MySQL, MariaDB and PostrgreSQL databases. If you want to use another DBI-supported database, you need to create tables in other ways.

Created tables should have following structure:

account (default name)

id: integer, primary key, auto-increment

username: varchar(255), NOT NULL, unique

password: varchar(255), NOT NULL

email: varchar(255), NOT NULL, unique

create_time: datetime, NOT NULL

update_time: datetime, NOT NULL

reset_code (default name)

id: integer, primary key, auto-increment

user_id: integer, NOT NULL

reset_code: varchar(10), NOT NULL

used: tinyint, NOT NULL

create_time: datetime, NOT NULL

update_time: datetime, NOT NULL

logs (default name, optional)

id: integer, primary key, auto-increment

time: datetime, NOT NULL

session: varchar(255), NOT NULL

direction: varchar(255), NOT NULL

type: varchar(255), NOT NULL

note: varchar(255)

See also

Examples

library(shiny.reglog)

# create a tenporary SQLite database
conn <- DBI::dbConnect(
  RSQLite::SQLite(),
  dbname = ":memory:"
)

# mockup user data
user_data <- 
  data.frame(username = c("Whatever", "Hanuka", "Helsinki", "How_come"), 
             password = c("&f5*MSYj^niDt=V'3.[dyEX.C/", "%}&B[fs\\}5PKE@,*+V\\tx9\"at]", 
                    "35z*ofW\\'G_8,@vCC`]~?e$Jm%", "s:;r_eLn?-D6;oA-=\"^R(-Ew<x"), 
             email = c("what@mil.com", "hehe@soso.so", "nider@what.no", "crazzz@simpsy.com"))

# create the tables and input the data (hashing the passwords in the process)
DBI_tables_create(conn = conn,
                  user_data = user_data,
                  hash_passwords = TRUE,
                  verbose = FALSE)
#> $account
#> $account$table_name
#> [1] "account"
#> 
#> $account$result
#> [1] TRUE
#> 
#> $account$data_import
#> [1] 4
#> 
#> 
#> $reset_code
#> $reset_code$table_name
#> [1] "reset_code"
#> 
#> $reset_code$result
#> [1] TRUE
#> 
#> 

# check generater tables
DBI::dbListTables(conn = conn)
#> [1] "account"    "reset_code"

# check the "user" table for user data
DBI::dbReadTable(conn = conn,
                 "account")
#>   id username
#> 1  1 Whatever
#> 2  2   Hanuka
#> 3  3 Helsinki
#> 4  4 How_come
#>                                                                                                                           password
#> 1 c2NyeXB0ABAAAAAIAAAAAcpVBCQmjLI6Tv44E+M1KLJb1HdMQYdUKUU9czEs+PL1Z6pxZBvt+9BrRY3HjllBZzNS/bCsqPbo153HCnJVcWiBuzCkTPoDx57uZzDJOOsw
#> 2 c2NyeXB0ABAAAAAIAAAAAfYPoM5qiOdm71smtcIOiPS3mFhaeV+1UTvfbG8PEjjbbg+yU7c0HLDFKqMw+zuJEumjg7rhfVJF7Geq/u3Cx9Zbh7kHEAXqxGaZoK03luN9
#> 3 c2NyeXB0ABAAAAAIAAAAAbgfE5OeJlmkHWfYoU78EXj/7fmYBW2gVJ9KiC6+0w6PbOdenkm4KPUgIBcGyc+zwUHDWgpaK+acRBwqgjDQVGLRsRdh3zD8aBkSLCGzjtZ6
#> 4 c2NyeXB0ABAAAAAIAAAAAYyOAm2wJcpkMOlMpSd1jb7EuYORIbPKkDsyVMHVniw1WcCkCqsGb3Sqyls+kvFndW5V6D7OhZMZ8AHoz6vkkLnMdrfQfj1bT4T1GJWMYMI5
#>               email             create_time             update_time
#> 1      what@mil.com 2022-03-24 17:40:49.324 2022-03-24 17:40:49.324
#> 2      hehe@soso.so 2022-03-24 17:40:49.324 2022-03-24 17:40:49.324
#> 3     nider@what.no 2022-03-24 17:40:49.324 2022-03-24 17:40:49.324
#> 4 crazzz@simpsy.com 2022-03-24 17:40:49.324 2022-03-24 17:40:49.324

# disconnect
DBI::dbDisconnect(conn = conn)