Connector to DBI-valid databases

Object of this class handles all connections for the RegLogServer object to the database. It is created to handle DBI compatible drivers. Provides methods than will be used by RegLogServer to get and send data.

See also

Super class

shiny.reglog::RegLogConnector -> RegLogDBIConnector

Methods

Public methods

RegLogDBIConnector$new()

RegLogDBIConnector$clone()

Inherited methods

shiny.reglog::RegLogConnector$get_logs()

shiny.reglog::RegLogConnector$resume()

shiny.reglog::RegLogConnector$suspend()

Method new()

Initialization of the object. Creates initial connection to the database.

Usage

RegLogDBIConnector$new(
  driver,
  ...,
  table_names = c("account", "reset_code", "logs"),
  custom_handlers = NULL
)

Arguments

driver

Call that specifies the driver to be used during all queries

...

other arguments used in DBI::dbConnect() call

table_names

character vector. Contains names of the tables in the database: first containing user data, second - reset codes information, third (optional) - logs from the object. For more info check documentation of DBI_database_create.

custom_handlers

named list of custom handler functions. Custom handler should take arguments: self and private - relating to the R6 object and message of class RegLogConnectorMessage. It should return return RegLogConnectorMessage object.

Returns

object of RegLogDBIConnector class

Method clone()

The objects of this class are cloneable with this method.

Usage

RegLogDBIConnector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.