Shiny UI module for registration box

It need to be used in conjuction with login_server() function and is suggested to be used alongside login_UI() and password_reset_UI() for full potential.

This function creates a UI div() element containing informations and input necessary for user to register new account. As it outputs a div() element, you can put it inside container of your choosing (be it some tabItem, fluidPage, fluidRow etc.)

Usage

register_UI(id = "login_system", lang = "en", custom_txts = NULL)

Arguments

id

the id of the module. Defaults to "login_system" for all of the modules contained within the package. If you plan to use serveral login systems inside your app or for any other reason need to change it, remember to keep consistent id for all elements of module.

lang

specifies the app used language. Accepts "en" or "pl". Defaults to "en"

custom_txts

named list containing customized texts. For more details, see documentation for 'reglog_txt'. Provided list can contain only elements used by this function, but it is recommended to provide the same list for every 'shiny.reglog' function

Value

'tags$div' containing all elements for registration procedure

See also